# docs/swagger.yaml
definitions:
  github_com_crazyfrankie_kube-ctl_internal_model_req.Base:
    properties:
      labels:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      name:
        type: string
      namespace:
        type: string
      restartPolicy:
        description: 'reboot strategy: Always | Never | On-Failure'
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.ConfigMapRefVolume:
    properties:
      name:
        type: string
      optional:
        type: boolean
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.Container:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        items:
          type: string
        type: array
      env:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVar'
        type: array
      envsFrom:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVarFromResource'
        type: array
      image:
        type: string
      imagePullPolicy:
        description: Always | IfNotPresent | Never
        type: string
      livenessProbe:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe'
        description: Survival probes
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerPort'
        type: array
      privileged:
        description: Whether to enable privileged mode (e.g. root)
        type: boolean
      readyProbe:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe'
        description: Readiness Probe
      resources:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Resource'
        description: Container application quota
      startUpProbe:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe'
        description: Start the probe
      tty:
        type: boolean
      volumeMounts:
        description: Mounted volumes
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.VolumeMount'
        type: array
      workingDir:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerPort:
    properties:
      containerPort:
        type: integer
      hostPort:
        type: integer
      name:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe:
    properties:
      command:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeCommand'
      enable:
        description: Whether to turn on the probe
        type: boolean
      failureThreshold:
        description: Threshold number of probes to consider a probe failure
        type: integer
      httpGet:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeHTTPGet'
      initialDelaySeconds:
        description: Initialize for a number of seconds before starting the probe
        type: integer
      periodSeconds:
        description: Probe after every few seconds
        type: integer
      successThreshold:
        description: Threshold number of probes to consider a probe successful
        type: integer
      tcpSocket:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeTcpSocket'
      timeoutSeconds:
        description: Probe timeout time
        type: integer
      type:
        description: 'Probe type: tcp | http | command'
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.DaemonSet:
    properties:
      labels:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      name:
        type: string
      namespace:
        type: string
      selector:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      template:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Pod'
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.Deployment:
    properties:
      labels:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      name:
        type: string
      namespace:
        type: string
      replicas:
        type: integer
      selector:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      template:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Pod'
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.DnsConfig:
    properties:
      nameservers:
        items:
          type: string
        type: array
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.DownwardAPIVolume:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DownwardAPIVolumeItem'
        type: array
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.DownwardAPIVolumeItem:
    properties:
      fieldPath:
        type: string
      path:
        description: pod internal path
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVar:
    properties:
      name:
        type: string
      refName:
        type: string
      type:
        description: configMap | secret | default(k/v)
        type: string
      value:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVarFromResource:
    properties:
      name:
        type: string
      prefix:
        type: string
      refType:
        description: configMap | secret
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.HostPathVolume:
    properties:
      path:
        description: host path
        type: string
      type:
        $ref: '#/definitions/v1.HostPathType'
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Item:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Network:
    properties:
      dnsConfig:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DnsConfig'
      dnsPolicy:
        type: string
      hostAliases:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      hostName:
        type: string
      hostNetwork:
        type: boolean
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.NodeAffinityTermExpressions:
    properties:
      key:
        type: string
      operator:
        $ref: '#/definitions/v1.NodeSelectorOperator'
      value:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.NodeScheduling:
    properties:
      nodeAffinity:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.NodeAffinityTermExpressions'
        type: array
      nodeName:
        type: string
      nodeSelector:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      type:
        description: nodeName | nodeSelector | nodeAffinity
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.PVCVolume:
    properties:
      claimName:
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Pod:
    properties:
      base:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Base'
        description: base definition info
      containers:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Container'
        type: array
      initContainers:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Container'
        type: array
      network:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Network'
      nodeScheduling:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.NodeScheduling'
      tolerations:
        description: pod toleration params
        items:
          $ref: '#/definitions/v1.Toleration'
        type: array
      volume:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Volume'
        type: array
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeCommand:
    properties:
      command:
        description: cat /test/test.txt
        items:
          type: string
        type: array
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeHTTPGet:
    properties:
      headers:
        description: http headers
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      host:
        description: If "", it's an in-pod request.
        type: string
      path:
        type: string
      port:
        type: integer
      scheme:
        description: http | https
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeTcpSocket:
    properties:
      host:
        type: string
      port:
        type: integer
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.Resource:
    properties:
      CPULimit:
        type: integer
      CPUReq:
        description: cpu m
        type: integer
      enable:
        description: Whether to configure container quotas
        type: boolean
      memory:
        description: memory mi
        type: integer
      memoryLimit:
        type: integer
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.SecretRefVolume:
    properties:
      name:
        type: string
      optional:
        type: boolean
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Volume:
    properties:
      PVCVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.PVCVolume'
      configMapRefVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ConfigMapRefVolume'
      downwardAPIVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DownwardAPIVolume'
      hostPathVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.HostPathVolume'
      name:
        type: string
      secretRefVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.SecretRefVolume'
      type:
        description: emptyDir | configMap | secret | hostPath | downward | pvc
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.VolumeMount:
    properties:
      mountName:
        type: string
      mountPath:
        description: Mounted volume -> path in the corresponding container
        type: string
      readOnly:
        description: Read-only or not
        type: boolean
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_resp.DaemonSet:
    properties:
      age:
        type: integer
      available:
        description: 表示可用的守护进程副本数，即已经就绪并且可以提供服务的守护进程副本数量
        type: integer
      current:
        description: 表示当前正在运行的守护进程副本数，即当前在集群中实际运行的守护进程副本数量。
        type: integer
      desired:
        description: 表示 DaemonSet 所需的守护进程副本数，即期望在集群中运行的守护进程副本数量。
        type: integer
      name:
        type: string
      namespace:
        type: string
      ready:
        description: 表示已经就绪的守护进程副本数，即当前运行且已准备好接收工作负载的守护进程副本数量。
        type: integer
      upToDate:
        description: 表示已经更新到最新版本的守护进程副本数，即已经与 DaemonSet 的定义保持一致的守护进程副本数量
        type: integer
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_resp.Deployment:
    properties:
      age:
        type: integer
      available:
        description: Available 字段表示 Deployment 中可用的 Pod 副本数
        type: integer
      name:
        type: string
      namespace:
        type: string
      ready:
        description: Ready 字段表示 Deployment 中正在运行的 Pod 副本的数量
        type: integer
      replicas:
        type: integer
      upToDate:
        description: UpToDate 字段表示与 Deployment 所期望的副本数相比，有多少个 Pod 副本是最新的
        type: integer
    type: object
...
  github_com_crazyfrankie_kube-ctl_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  v1.HostPathType:
    enum:
    - ""
    - DirectoryOrCreate
    - Directory
    - FileOrCreate
    - File
    - Socket
    - CharDevice
    - BlockDevice
    type: string
    x-enum-varnames:
    - HostPathUnset
    - HostPathDirectoryOrCreate
    - HostPathDirectory
    - HostPathFileOrCreate
    - HostPathFile
    - HostPathSocket
    - HostPathCharDev
    - HostPathBlockDev
...
  v1.NodeSelectorOperator:
    enum:
    - In
    - NotIn
    - Exists
    - DoesNotExist
    - Gt
    - Lt
    type: string
    x-enum-varnames:
    - NodeSelectorOpIn
    - NodeSelectorOpNotIn
    - NodeSelectorOpExists
    - NodeSelectorOpDoesNotExist
    - NodeSelectorOpGt
    - NodeSelectorOpLt
...
  v1.TaintEffect:
    enum:
    - NoSchedule
    - PreferNoSchedule
    - NoExecute
    type: string
    x-enum-varnames:
    - TaintEffectNoSchedule
    - TaintEffectPreferNoSchedule
    - TaintEffectNoExecute
  v1.Toleration:
    properties:
      effect:
        allOf:
        - $ref: '#/definitions/v1.TaintEffect'
        description: |-
          Effect indicates the taint effect to match. Empty means match all taint effects.
          When specified, allowed values are NoSchedule, PreferNoSchedule and NoExecute.
          +optional
      key:
        description: |-
          Key is the taint key that the toleration applies to. Empty means match all taint keys.
          If the key is empty, operator must be Exists; this combination means to match all values and all keys.
          +optional
        type: string
      operator:
        allOf:
        - $ref: '#/definitions/v1.TolerationOperator'
        description: |-
          Operator represents a key's relationship to the value.
          Valid operators are Exists and Equal. Defaults to Equal.
          Exists is equivalent to wildcard for value, so that a pod can
          tolerate all taints of a particular category.
          +optional
      tolerationSeconds:
        description: |-
          TolerationSeconds represents the period of time the toleration (which must be
          of effect NoExecute, otherwise this field is ignored) tolerates the taint. By default,
          it is not set, which means tolerate the taint forever (do not evict). Zero and
          negative values will be treated as 0 (evict immediately) by the system.
          +optional
        type: integer
      value:
        description: |-
          Value is the taint value the toleration matches to.
          If the operator is Exists, the value should be empty, otherwise just a regular string.
          +optional
        type: string
    type: object
  v1.TolerationOperator:
    enum:
    - Exists
    - Equal
    type: string
    x-enum-varnames:
    - TolerationOpExists
    - TolerationOpEqual
...
info:
  contact: {}
paths:
...
  /api/daemonset:
    delete:
      consumes:
      - application/json
      description: 删除指定命名空间下的指定DaemonSet
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 删除 DaemonSet
      tags:
      - DaemonSet 管理
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下指定DaemonSet的详细信息
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回DaemonSet的详细信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DaemonSet'
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取DaemonSet详情
      tags:
      - DaemonSet 管理
    post:
      consumes:
      - application/json
      description: 创建新的 DaemonSet 或更新已存在的 DaemonSet
      parameters:
      - description: DaemonSet 配置信息
        in: body
        name: pod
        required: true
        schema:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DaemonSet'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功，返回成功消息
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "400":
          description: 参数错误(code=20001)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 创建或更新 DaemonSet
      tags:
      - DaemonSet 管理
  /api/daemonset/list:
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下的所有DaemonSet列表
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 关键词
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回DaemonSet列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.DaemonSet'
                  type: array
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取DaemonSet列表
      tags:
      - DaemonSet 管理
  /api/deployment:
    delete:
      consumes:
      - application/json
      description: 删除指定命名空间下的指定Deployment
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Deployment 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 删除 Deployment
      tags:
      - Deployment 管理
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下指定Deployment的详细信息
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Deployment 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回Deployment的详细信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Deployment'
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取Deployment详情
      tags:
      - Deployment 管理
    post:
      consumes:
      - application/json
      description: 创建新的 Deployment 或更新已存在的 Deployment
      parameters:
      - description: Deployment 配置信息
        in: body
        name: pod
        required: true
        schema:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Deployment'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功，返回成功消息
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "400":
          description: 参数错误(code=20001)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 创建或更新 Deployment
      tags:
      - Deployment 管理
  /api/deployment/list:
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下的所有Deployment列表
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 关键词
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回Deployment列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.Deployment'
                  type: array
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取Deployment列表
      tags:
      - Deployment 管理
...
swagger: "2.0"